+++ iUnits/iUnits/Model/Categories.swift
@@ -1,0 +1,31 @@
+//
+//  Categories.swift
+//  iUnits
+//
+//  Created by Israel Manzo on 9/12/23.
+//
+
+import Foundation
+
+enum Category: String, CaseIterable {
+    case length
+    case measurement
+    case speed
+    case area
+    case volumen
+    case mass
+    case temperature
+    case  pressure
+}
+
+enum Length: String, CaseIterable {
+    case millimeters
+    case centimeters
+    case meters
+    case kilometers
+    case miles
+    case inches
+    case feet
+    case yards
+}
+
